🚚rename: api, component 파일 정리 방식 수정 #80
