+++ client/src/components/Login/LoginSlider/style.ts
